Add 32 bit generator support and glibc random()
Add support for generators that can generate 32 bits at a time, and for
those that can only generate 32 bits at a time. For the latter, if 64
bits are required the generator will by default be called twice and the
results concatenated. As an example, a copy of the (not very
statistically strong) glibc random() function has been implemented.