Add some basic sanity checks for generators
Sanity_Checks and Sanity_Checks32 are generic procedures that carry out
some basic checks that generators can be reset, can be seeded, are not
simply constant etc. They should be applicable to all generators and
catch basic errors like forgetting to write the state back to the
generator after updating it.