Generate basic uniform, normal variates
Start on routines to generate basic variates. The code is written as a
generic with parameters for floating-point type and PRNG type. This
allows optimal performance in most simulation tasks where there will be
no run-time switching of generators. The dispatching meta-PRNG can be
used if this is required, at the expense of slightly more messy code.