Refactor distributions as a stand-alone hierarchy
Distributions should not really be a generic type extension to the PRNG
as this makes little sense. It also greatly complicates using a single
PRNG as the source for several different variates in a simulation. It
is probably best to separate them out and require the appropriate PRNG
to be specified directly whenever a variate is required.