Require all PRNG to support 32 and 64 bit output
All PRNG are now required to support 32 and 64 bit output by the PRNG
interface. However there are PRNG_32Only and PRNG_64Only abstract types
which implement calling a 32 bit PRNG twice to produce the 64 bit
output, and masking off the top 32 bits from a 64 bit PRNG to produce a
32 bit output. While more consistent, this does often require subtype
indications to show which Generate function is meant - this are may need
some more thought.