Add support for tests based on binned variates
Started to add support for testing the actual distribution of variates
between a set of bins versus the expected distribution. For example,
the (t,l) equidistribution of the raw output of a PRNG could be tested,
or the quality of the combination of a PRNG and a routine to create
normal variates.