Refactor test framework using multiple inheritance
Rather than having separate class hierarchies of tests and test results
and dealing with the problems of the lack of any link between a test
type and the associate test result type, they are now combined. MI is
used so different sorts of tests can all have a basic reporting
interface while having different methods for creating and running the
tests.