Correct EquiDist test for masking of PRNG output
Where only part of the PRNG output is valid the remaining bits should be
zeroed out with a mask, or else invalid access will occur. This was not
previously identified as all of the narrow-output PRNG that were
previously tested happened to produce all zeros in the unused part of
the output.