Better accommodate PRNG of different widths
Two new primitive operations have been added to the PRNG root type.
Width is a function that gives the usable width of the output, which may
not always be the full 32- or 64- bits. Generate_Padded generates a
64-bit output by padding unused bits with 0, as opposed to the regular
Generate with 64-bit output where PRNG with shorter outputs are expected
to concatenate multiple outputs until they reach the correct width.