sample_output can now seed using a 1-element array
The sample_output example can now seed the generator using a 1-element
array rather than a single value. This is useful for some PRNG such as
the Mersenne Twister series where there is a canonical way of seeding
from arrays that is different from the canonical way of seeding from
scalar values.