Add consistency checking to speedtest
Add some consistency checking to the speedtest example by splitting the
test runs down into subsets and using the median of the subset results
as the final result. The top and bottom subset results are compared with
the median, and if outside a tolerance value the whole set is run again
until they are (or an attempt limit is reached).

There is still an issue with test results getting faster when the same
generator is tested repeatedly - this may be some sort of 'warm-up'
effect relating to instruction branch prediction.