sample_output can now be seeded from an array
The parse_args package now supports integer array arguments so the
sample_output example can now initialise PRNG from an array. This is
mainly useful where there is a canonical example of how to do this,
such as with the Mersenne Twister family, as it allows easy comparison
with the author's example code. The test_mt example is now unnecessary.