Fix loop terminating condition in Chi2_CDF
Put in a proper loop terminating condition in the Chi2_CDF function,
based on an upper bound for the sum of the remaining terms. One of the
unit tests was adjusted as it was too strict for the new default
epsilon. Some variables were renamed to match the notes more closely.