Correct bit counter test
The overall pass-fail result should use an adjusted p-value when
considering if any of the individual bit tests failed. Effectively,
requiring all the individual bit tests to pass in order to give an
overall pass was vulnerable to spurious failures as it amounted to
repeating an experiment many times.

Bit counter test result was also only labelling failures b-d rather than
a-d as expected.